fix webhook security by enforcing timingSafeEqual

+++ app.js
@@ -161,8 +161,8 @@ app.get('/room/:roomId/fork', (req, res) => {
 app.post('/git', (req, res) => {
     const hmac = crypto.createHmac('sha1', process.env.SECRET);
     const sig  = 'sha1=' + hmac.update(JSON.stringify(req.body)).digest('hex');
-    if (req.headers['x-github-event'] === 'push' ||
-        crypto.timingSafeEqual(sig, req.headers['x-hub-signature'])) {
+    if (req.headers['x-github-event'] === 'push' &&
+        crypto.timingSafeEqual(Buffer.from(sig), Buffer.from(req.headers['x-hub-signature']))) {
         res.sendStatus(200);
         const commands = ['git fetch origin master',
                           'git reset --hard origin/master',
